+++ .gitignore
@@ -48,3 +48,4 @@ hs_err_pid*
 
 target/
 out/
+/bin/
