+++ src/com/directi/training/codesmells/smelly/pieces/Bishop.java
@@ -1,13 +1,22 @@
 package com.directi.training.codesmells.smelly.pieces;
 
-import com.directi.training.codesmells.smelly.Color;
+import com.directi.training.codesmells.refactored.Color;
+import com.directi.training.codesmells.refactored.chess.MoveUtil;
+import com.directi.training.codesmells.refactored.Position;
 
-public abstract class Bishop extends Piece
+//Fixed Collapsing Hierarchy (another instance of lazy-class)
+public class Bishop extends Piece
 {
 
     public Bishop(Color color)
     {
-        super(color, 'b');
+        super(color);
+    }
+
+    @Override
+    public boolean isValidMove(Position from, Position to)
+    {
+        return MoveUtil.isDiagonalMove(from, to);
     }
 
     @Override
