+++ src/com/directi/training/codesmells/smelly/pieces/King.java
@@ -1,12 +1,19 @@
 package com.directi.training.codesmells.smelly.pieces;
 
-import com.directi.training.codesmells.smelly.Color;
+import com.directi.training.codesmells.refactored.Color;
+import com.directi.training.codesmells.refactored.Position;
 
 public class King extends Piece
 {
     public King(Color color)
     {
-        super(color, 'K');
+        super(color);
+    }
+
+    @Override
+    public boolean isValidMove(Position from, Position to)
+    {
+        return (Math.abs(from.getRow() - to.getRow()) == 1) && (Math.abs(from.getColumn() - to.getColumn()) == 1);
     }
 
     @Override
