+++ src/com/directi/training/codesmells/smelly/pieces/Knight.java
@@ -1,12 +1,21 @@
 package com.directi.training.codesmells.smelly.pieces;
 
-import com.directi.training.codesmells.smelly.Color;
+import com.directi.training.codesmells.refactored.Color;
+import com.directi.training.codesmells.refactored.Position;
 
-public abstract class Knight extends Piece
+//Fixed Collapsing Hierarchy (another instance of lazy-class)
+public class Knight extends Piece
 {
     public Knight(Color color)
     {
-        super(color, 'k');
+        super(color);
+    }
+
+    public boolean isValidMove(Position from, Position to)
+    {
+        int columnDiff = Math.abs(to.getColumn() - from.getColumn());
+        int rowDiff = Math.abs(to.getRow() - from.getRow());
+        return (columnDiff == 2 && rowDiff == 1) || (columnDiff == 1 && rowDiff == 2);
     }
 
     @Override
