--- src/com/directi/training/codesmells/smelly/pieces/LeftKnight.java
@@ -1,11 +1,0 @@
-package com.directi.training.codesmells.smelly.pieces;
-
-import com.directi.training.codesmells.smelly.Color;
-
-public class LeftKnight extends Knight
-{
-    public LeftKnight(Color color)
-    {
-        super(color);
-    }
-}
