+++ src/com/directi/training/codesmells/smelly/pieces/Queen.java
@@ -1,12 +1,19 @@
 package com.directi.training.codesmells.smelly.pieces;
 
-import com.directi.training.codesmells.smelly.Color;
+import com.directi.training.codesmells.refactored.Color;
+import com.directi.training.codesmells.refactored.chess.MoveUtil;
+import com.directi.training.codesmells.refactored.Position;
 
 public class Queen extends Piece
 {
     public Queen(Color color)
     {
-        super(color, 'q');
+        super(color);
+    }
+
+    public boolean isValidMove(Position from, Position to)
+    {
+        return MoveUtil.isStraightLineMove(from, to);
     }
 
     @Override
